pyproject.toml
[tool.poetry]
name = "book-store-demoqa-web-api-tests"
version = "0.1.0"
description = ""
authors = ["natalia-s-belova <natalya.s.belova@gmail.com>"]
readme = "README.md"
packages = [{include = "book_store_demoqa_web_api_tests"}]

[tool.poetry.dependencies]
python = "^3.10"
selene = ">=2.0.0rc4,<2.1.0"
allure-pytest = "^2.13.2"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
curlify = "^2.2.1"
jsonschema = "^4.19.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
